+++ CHANGELOG.md
@@ -19,6 +19,7 @@ v 0.9.13
 - Cleaned up types to use a single MyTypes type
 - added prendyOptions to prendyAssets so there's less options to pass around
 - added helper and types for defining dolls and characters
+- Removed imports of other types from declarations.ts
 
 v 0.9.12